+++ CHANGES.md
@@ -4,3 +4,4 @@
 
 - Proof-of-concept and documentation, with software tests
 - Add program `node-blue setup`, to install JavaScript prerequisites
+- [fm] Add Node-BLUE flow manager and provider